Simplified consumer configuration by unifying existing ones into a global one. Pre checking of existing topics in current test.
