+++ src/main/java/com/umantis/poc/Consumer.java
@@ -1,28 +1,80 @@
 package com.umantis.poc;
 
+import com.umantis.poc.exponentialbackoff.RandomException;
+import org.apache.kafka.clients.consumer.ConsumerRecord;
+import org.apache.kafka.common.TopicPartition;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.kafka.annotation.KafkaListener;
+import org.springframework.kafka.listener.AcknowledgingMessageListener;
+import org.springframework.kafka.listener.ConsumerSeekAware;
+import org.springframework.kafka.support.Acknowledgment;
+import java.util.Map;
 import java.util.concurrent.CountDownLatch;
 
 /**
- * Basic spring kafka consumer
+ * Consumer implementation for exponential backoff message retry
  *
  * @author David Espinosa.
  */
-public class Consumer {
+public class Consumer implements AcknowledgingMessageListener<Integer, String>, ConsumerSeekAware {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(Consumer.class);
 
-    private CountDownLatch latch = new CountDownLatch(1);
+    private ConsumerSeekCallback consumerSeekCallback;
 
-    @KafkaListener(id = "consumer", topics = "${kafka.topic}", containerFactory = "kafkaListenerContainerFactory")
-    public void receive(String message) {
-        LOGGER.info("received message= "+message);
-        latch.countDown();
+    private CountDownLatch incorrectMessageLatch = new CountDownLatch(2);
+    private CountDownLatch correctMessageLatch = new CountDownLatch(1);
+
+    @Override
+    public void registerSeekCallback(final ConsumerSeekCallback consumerSeekCallback) {
+        this.consumerSeekCallback = consumerSeekCallback;
     }
 
-    public CountDownLatch latch() {
-        return latch;
+    @Override
+    public void onPartitionsAssigned(final Map<TopicPartition, Long> map, final ConsumerSeekCallback consumerSeekCallback) {
+
+    }
+
+    @Override
+    public void onIdleContainer(final Map<TopicPartition, Long> map, final ConsumerSeekCallback consumerSeekCallback) {
+
+    }
+
+    @KafkaListener(id = "consumer", topics = "${kafka.topic}")
+    public void receive(String message) {
+        LOGGER.info("received message= "+message);
+        correctMessageLatch.countDown();
+    }
+
+    @Override
+    @KafkaListener(id = "seeker", topics = "${kafka.seeker_topic}")
+    public void onMessage(final ConsumerRecord<Integer, String> consumerRecord, final Acknowledgment acknowledgment) {
+
+        try {
+            String value = (String) consumerRecord.value();
+            if (value.contains("NOT")) {
+                boolean emulateError = (incorrectMessageLatch.getCount() == 2);
+                incorrectMessageLatch.countDown();
+                if (emulateError) {
+                    throw new RandomException("Random Exception to try message re-processing!");
+                }
+            } else {
+                acknowledgment.acknowledge();
+                correctMessageLatch.countDown();
+            }
+        } catch (RandomException e) {
+            consumerSeekCallback.seek(consumerRecord.topic(), consumerRecord.partition(), consumerRecord.offset());
+            LOGGER.error("Error processing message with offset: " + consumerRecord.offset() + " from topic: " + consumerRecord.topic());
+        }
+    }
+
+    public CountDownLatch getIncorrectMessageLatch() {
+        return incorrectMessageLatch;
+    }
+
+    public CountDownLatch getLatch() {
+        return correctMessageLatch;
     }
 }
