+++ src/main/resources/application.properties
@@ -3,8 +3,7 @@ kafka.topic=poc-topic
 kafka.non_existing_topic=non-existing-topic
 kafka.retention_topic=retention-topic
 kafka.seeker_topic=seeker-topic
+consumer.grouip=normal_consumer
 zookeeper.server=localhost:2181
 zookeeper.session_timeout=10000
 zookeeper.connection_timeout=10000
-consumer.grouip=normal_consumer
-seeker_consumer.grouip=seeker_consumer
