+++ src/test/java/com/umantis/poc/SpringKafkaApplicationTest.java
@@ -1,23 +1,15 @@
 package com.umantis.poc;
 
-import java.util.concurrent.TimeUnit;
-
+import com.umantis.poc.admin.KafkaAdminUtils;
 import org.assertj.core.api.Assertions;
 import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.ClassRule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.kafka.config.KafkaListenerEndpointRegistry;
-import org.springframework.kafka.listener.MessageListenerContainer;
-import org.springframework.kafka.test.rule.KafkaEmbedded;
-import org.springframework.kafka.test.utils.ContainerTestUtils;
 import org.springframework.test.context.junit4.SpringRunner;
-
-import com.umantis.poc.admin.KafkaAdminUtils;
+import java.util.concurrent.TimeUnit;
 
 /**
  * @author David Espinosa.
@@ -35,17 +27,17 @@ public class SpringKafkaApplicationTest {
     @Autowired
     public KafkaAdminUtils kafkaAdminService;
 
-    private static String TOPIC = "kafka-poc";
+    private static String TOPIC;
 
     @Value("${kafka.topic}")
     public void setTopic(String topic) {
         TOPIC = topic;
     }
-    
+
     @Before
     public void setup() {
-        if (kafkaAdminService.topicExists(TOPIC)) {
-            TOPIC = TOPIC + System.currentTimeMillis();
+        if (!kafkaAdminService.topicExists(TOPIC)) {
+            kafkaAdminService.createTopic(TOPIC, -1);
         }
     }
 
@@ -53,7 +45,7 @@ public class SpringKafkaApplicationTest {
     public void testReceive() throws Exception {
         producer.send(TOPIC, "Hello kafka");
 
-        consumer.latch().await(10000, TimeUnit.MILLISECONDS);
-        Assertions.assertThat(consumer.latch().getCount()).isEqualTo(0);
+        consumer.getLatch().await(10000, TimeUnit.MILLISECONDS);
+        Assertions.assertThat(consumer.getLatch().getCount()).isEqualTo(0);
     }
 }
